Use "token invalidate", skip when "--no-wait"

+++ phase/install_workers.go
@@ -77,11 +77,13 @@ func (p *InstallWorkers) Run() error {
 		return err
 	}
 
-	defer func() {
-		if err := p.leader.Exec(p.leader.Configurer.K0sCmdf("invalidate %s", token), exec.RedactString(token)); err != nil {
-			log.Warnf("%s: failed to invalidate the join token", p.leader)
-		}
-	}()
+	if !NoWait {
+		defer func() {
+			if err := p.leader.Exec(p.leader.Configurer.K0sCmdf("token invalidate %s", token), exec.RedactString(token)); err != nil {
+				log.Warnf("%s: failed to invalidate the join token", p.leader)
+			}
+		}()
+	}
 
 	return p.hosts.ParallelEach(func(h *cluster.Host) error {
 		log.Infof("%s: writing join token", h)
@@ -89,11 +91,13 @@ func (p *InstallWorkers) Run() error {
 			return err
 		}
 
-		defer func() {
-			if err := h.Configurer.DeleteFile(h, h.K0sJoinTokenPath()); err != nil {
-				log.Warnf("%s: failed to clean up the join token file at %s", h, h.K0sJoinTokenPath())
-			}
-		}()
+		if !NoWait {
+			defer func() {
+				if err := h.Configurer.DeleteFile(h, h.K0sJoinTokenPath()); err != nil {
+					log.Warnf("%s: failed to clean up the join token file at %s", h, h.K0sJoinTokenPath())
+				}
+			}()
+		}
 
 		if sp, err := h.Configurer.ServiceScriptPath(h, h.K0sServiceName()); err == nil {
 			if h.Configurer.ServiceIsRunning(h, h.K0sServiceName()) {
